Change enrollment input args to new format
