+++ drivers/hmis/app/graphql/types/hmis_schema/enrollment_household_member_input.rb
@@ -1,0 +1,8 @@
+module Types
+  class HmisSchema::EnrollmentHouseholdMemberInput < BaseInputObject
+    description 'HMIS Enrollment household member input'
+
+    argument :id, ID, required: true
+    argument :relationship_to_ho_h, HmisSchema::Enums::RelationshipToHoH, required: true
+  end
+end
